# node: optimize CBlockIndexWorkComparator