## Summary

This PR optimizes `CBlockIndexWorkComparator::operator()` by removing 4 redundant branches.

The previous implementation used multiple separate comparisons with explicit branches for greater-than and less-than cases, resulting in unnecessary code paths.

The new implementation consolidates comparisons into single inequality checks and reduces complexity while preserving its original behavior. This change is particularly beneficial for loading blocks from files and reindexing.

## Benchmarks

```shell
taskset -c 1 ./bin/bench_bitcoin --filter="(CheckBlockIndex|LoadExternalBlockFile|WalletAvailableCoins|WalletBalanceDirty|WalletCreatePlain|BlockToJsonVerboseWrite|WalletCreateTxUseOnlyPresetInputs|WalletIsMineMigratedDescriptors|WalletIsMineDescriptors)" -output-csv=bench_old.csv --min-time=30000
```

> Before:

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|       26,557,419.20 |               37.65 |    0.1% |     32.85 | `BlockToJsonVerboseWrite`
|          129,988.82 |            7,692.97 |    0.0% |     33.02 | `CheckBlockIndex`
|       21,661,396.96 |               46.17 |    0.5% |     32.37 | `LoadExternalBlockFile`
|              121.88 |        8,204,733.84 |    0.0% |     33.02 | `WalletIsMineDescriptors`

> After:

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|       27,930,130.95 |               35.80 |    0.1% |     32.96 | `BlockToJsonVerboseWrite`
|          115,346.65 |            8,669.52 |    0.0% |     33.00 | `CheckBlockIndex`
|       20,389,679.85 |               49.04 |    0.4% |     31.76 | `LoadExternalBlockFile`
|              135.01 |        7,406,957.83 |    0.0% |     33.01 | `WalletIsMineDescriptors`


Compared to master:

* `CheckBlockIndex` **+12.7%** faster
* `LoadExternalBlockFile` **+6.2%** faster
* `BlockToJsonVerboseWrite` **-4.9%** slower
* `WalletIsMineDescriptors` **-9.7%** slower

Arguably the most impacting improvement is LoadExternalBlockFile, with nearly 6% faster execution, which would significantly reduce block-loading times in typical workloads.