@Raimo33 2025-09-07
The reason `BlockToJsonVerboseWrite` and `WalletIsMineDescriptors` are slower is due to the fact that they benefit from not computing a comparison twice.

For them specifically, this would be better:
```diff
// First sort by most total work, ...
- if (pa->nChainWork != pb->nChainWork) {
-     return pa->nChainWork < pb->nChainWork;
- }
+ if (pa->nChainWork > pb->nChainWork) return false;
+ if (pa->nChainWork < pb->nChainWork) return true;
```

as in the best case scenario it only computes a single comparison. But I argue that a **+12.7%** improvement on `CheckBlockIndex` is way more significant